+++ packages/expo-app-intents/plugin/src/index.ts
@@ -2,3 +2,4 @@ import { Props } from './withAppIntents';
 
 export default (props: Props = {}): [string, Props] => ['expo-app-intents', props];
 export { default as withAppIntents } from './withAppIntents';
+export type { IntentConfig } from './generatedIntents';